+++ instrumentation/runner/src/main/kotlin/de/mannodermaus/junit5/internal/runners/AndroidJUnit5.kt
@@ -5,6 +5,7 @@ import android.util.Log
 import androidx.annotation.VisibleForTesting
 import de.mannodermaus.junit5.internal.LOG_TAG
 import de.mannodermaus.junit5.internal.LibcoreAccess
+import de.mannodermaus.junit5.internal.runners.notification.ParallelRunNotifier
 import org.junit.platform.launcher.core.LauncherFactory
 import org.junit.runner.Runner
 import org.junit.runner.notification.RunNotifier
@@ -12,25 +13,19 @@ import org.junit.runner.notification.RunNotifier
 /**
  * JUnit Runner implementation using the JUnit Platform as its backbone.
  * Serves as an intermediate solution to writing JUnit 5-based instrumentation tests
- * until official support arrives for this. This is in Java because we require access to package-private data,
- * and Kotlin is more strict about that: https://youtrack.jetbrains.com/issue/KT-15315
- *
- *
- * Replacement For:
- * AndroidJUnit4
+ * until official support arrives for this.
  *
  * @see org.junit.platform.runner.JUnitPlatform
  */
 @SuppressLint("NewApi")
 @VisibleForTesting(otherwise = VisibleForTesting.PACKAGE_PRIVATE)
-internal class AndroidJUnit5
-@JvmOverloads constructor(
+internal class AndroidJUnit5(
     private val testClass: Class<*>,
-    private val runnerParams: AndroidJUnit5RunnerParams = createRunnerParams(testClass)
+    private val runnerParams: AndroidJUnit5RunnerParams = createRunnerParams(testClass),
 ) : Runner() {
 
     private val launcher = LauncherFactory.create()
-    private val testTree = generateTestTree(runnerParams)
+    private val testTree by lazy { generateTestTree(runnerParams) }
 
     override fun getDescription() =
         testTree.suiteDescription
@@ -41,7 +36,10 @@ internal class AndroidJUnit5
         registerSystemProperties()
 
         // Finally, launch the test plan on the JUnit Platform
-        launcher.execute(testTree.testPlan, AndroidJUnitPlatformRunnerListener(testTree, notifier))
+        launcher.execute(
+            testTree.testPlan,
+            AndroidJUnitPlatformRunnerListener(testTree, createNotifier(notifier)),
+        )
     }
 
     /* Private */
@@ -66,9 +64,19 @@ internal class AndroidJUnit5
         }
     }
 
-    private fun generateTestTree(params: AndroidJUnit5RunnerParams): AndroidJUnitPlatformTestTree {
-        val discoveryRequest = params.createDiscoveryRequest()
-        val testPlan = launcher.discover(discoveryRequest)
-        return AndroidJUnitPlatformTestTree(testPlan, testClass, params.isIsolatedMethodRun())
-    }
+    private fun generateTestTree(params: AndroidJUnit5RunnerParams) =
+        AndroidJUnitPlatformTestTree(
+            testPlan = launcher.discover(params.createDiscoveryRequest()),
+            testClass = testClass,
+            isIsolatedMethodRun = params.isIsolatedMethodRun,
+            isParallelExecutionEnabled = params.isParallelExecutionEnabled,
+        )
+
+    private fun createNotifier(nextNotifier: RunNotifier) =
+        if (testTree.isParallelExecutionEnabled) {
+            // Wrap the default notifier with a special handler for parallel test execution
+            ParallelRunNotifier(nextNotifier)
+        } else {
+            nextNotifier
+        }
 }