+++ instrumentation/sample/src/androidTest/kotlin/de/mannodermaus/sample/TestRunningOnJUnit4.kt
@@ -8,4 +8,24 @@ class TestRunningOnJUnit4 {
   fun junit4() {
     Assert.assertEquals(4, 2 + 2)
   }
+
+  @Test
+  fun junit4_2() {
+    Assert.assertEquals(4, 2 + 2)
+  }
+
+  @Test
+  fun junit4_3() {
+    Assert.assertEquals(4, 2 + 2)
+  }
+
+  @Test
+  fun junit4_4() {
+    Assert.assertEquals(4, 2 + 2)
+  }
+
+  @Test
+  fun junit4_5() {
+    Assert.assertEquals(4, 2 + 2)
+  }
 }