+++ instrumentation/CHANGELOG.md
@@ -7,6 +7,7 @@ Change Log
 - Add support for inherited tests (#288)
 - Only autoconfigure JUnit 5 for instrumentation tests when the user explicitly adds junit-jupiter-api as a dependency
 - Prevent noisy logs in Logcat complaining about unresolvable annotation classes (#306)
+- Add support for parallel execution of non-UI instrumentation tests (#295)
 
 ## 1.3.0 (2021-09-17)